pyproject.toml
[tool.poetry]
name = "Guidance for Generative AI Deployments using Amazon SageMaker JumpStart"
version = "1.0.0"
description = ""
authors = [
    "Constantin Gonzalez <glez@amazon.com>",
    "Robert Hanuschke <robhan@amazon.com>",
]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "^2.85.0"
aws-cdk-aws-lambda-python-alpha = "^2.85.0a0"
tomli = "^2.0.1"
cdk-nag = "^2.27.53"
sagemaker = "^2.168.0"
boto3 = "^1.26.162"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
pre-commit = "^3.3.3"
aws-lambda-powertools = "^2.18.0"
botocore = "^1.29.162"
boto3-stubs = { extras = [
    "s3",
    "sagemaker-runtime",
    "ssm",
], version = "^1.26.162" }
pillow = "^9.5.0"
pipreqs = "^0.4.13"
pip-licenses = "^4.3.2"
semgrep = "^1.29.0"
bandit = "^1.7.5"
mypy = "^1.3.0"
black = "^23.3.0"

[tool.ruff]
ignore = ["E501"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
